requirements.txt
playwright==1.46.0
pydantic==2.8.2
sqlalchemy~=2.0.30
redis~=5.0.5
hiredis==3.0.0
pytz~=2024.1
python-dotenv~=1.0.1
PyMySQL==1.1.1
DrissionPage==4.0.5.6
twitter-api-client==0.10.22